moved code into a worker which can be imported by ai and locally running scripts
